requirements.txt
kafka-python==2.0.2
faker==19.6.2
pyspark==3.4.1
xgboost==2.0.0
scikit-learn==1.3.1
fastapi==0.104.0
uvicorn==0.24.0
redis==5.0.1
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.24.4
joblib==1.3.2
python-dotenv==1.0.0
httpx==0.25.0
pytest==7.4.3
